requirements.txt
flask
yt-dlp
requests
google-generativeai
gunicorn
archiver
